chore(ETH): linting ETH hello world contract

+++ ethereum/hello_world/HelloWrold.sol
@@ -2,17 +2,17 @@
 pragma solidity 0.8.19;
 
 contract HelloWorld {
-    mapping(string => uint) public nameToPoints;
+    mapping(string => uint256) public nameToPoints;
 
-    function setPoint(string memory _Name, uint _point) public {
+    function setPoint(string memory _Name, uint256 _point) public {
         nameToPoints[_Name] = _point;
     }
 
-    function name() public pure returns(string memory) {
+    function name() public pure returns (string memory) {
         return "Hello, World!";
     }
 
-    function lookup(string memory _Name)  public  view returns(uint) {
+    function lookup(string memory _Name) public view returns (uint256) {
         return nameToPoints[_Name];
     }
 }
